requirements.txt
flask
joblib
pandas
numpy
scikit-learn
sqlalchemy
psycopg2-binary
python-dotenv
kagglehub
